requirements.txt
altgraph==0.17.2
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
auto-py-to-exe==2.18.0
autopep8==1.5
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
bottle==0.12.19
bottle-websocket==0.2.9
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.49.0
django-crispy-forms==1.14.0
django-filter==21.1
django-heroku==0.3.1
django-registration==3.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
duplicity==0.8.12.0
Eel==0.12.4
entrypoints==0.3
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
mysql-connector-python==8.0.28
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pep8==1.7.1
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.1
protobuf==3.6.1
psycopg2==2.8.4
psycopg2-binary==2.9.3
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
PyGObject==3.36.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.2.4
systemd-python==234
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.13.2
wadllib==1.3.3
whichcraft==0.6.1
whitenoise==6.0.0
xkit==0.0.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
